// src/client/algorithm1/model/course.ts
/**
 * algorithm-1
 * This is the Algorithm-1 API endpoint specs for SENG499 Capstone Company 3 & 4.
 *
 * The version of the OpenAPI document: 2.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Assignment } from './assignment';
import { Professor } from './professor';

export class Course {
  'courseNumber': string;
  'subject': string;
  'squenceNumber'?: string;
  'courseTitle': string;
  'assignment'?: Array<Assignment>;
  'prof'?: Array<Professor>;

  static discriminator: string | undefined = undefined;

  static attributeTypeMap: Array<{
    name: string;
    baseName: string;
    type: string;
  }> = [
    {
      name: 'courseNumber',
      baseName: 'CourseNumber',
      type: 'string',
    },
    {
      name: 'subject',
      baseName: 'Subject',
      type: 'string',
    },
    {
      name: 'squenceNumber',
      baseName: 'SquenceNumber',
      type: 'string',
    },
    {
      name: 'courseTitle',
      baseName: 'CourseTitle',
      type: 'string',
    },
    {
      name: 'assignment',
      baseName: 'Assignment',
      type: 'Array<Assignment>',
    },
    {
      name: 'prof',
      baseName: 'Prof',
      type: 'Array<Professor>',
    },
  ];

  static getAttributeTypeMap() {
    return Course.attributeTypeMap;
  }
}